+++ packages/dito/app/resources/allRoutes.ts
@@ -1,15 +1,5 @@
-import { SVGProps } from "react";
-import routes from "./allRoutes";
 import { precheck } from "./content";
-import { PATH_PRECHECK } from "./staticRoutes";
-
-export type Route = {
-  url: string;
-  parent?: string;
-} & (
-  | { title: string }
-  | { Icon: React.ComponentType<SVGProps<SVGSVGElement>> }
-);
+import routes, { PATH_PRECHECK, type Route } from "./staticRoutes";
 
 const allRoutes: Route[] = [
   ...routes,
